Токен Сохранение данных при редактировании профиля

+++ app/src/main/java/com/example/myvkclient/presentation/screen/profile/feed/BaseMessage.kt
@@ -1,3 +1,3 @@
 package com.example.myvkclient.presentation.screen.profile.feed
 
-abstract class BaseMessage(val id: Int)
+abstract class BaseMessage(val id: Long)
